Version Final Gerarquia
