+++ src/main/java/com/tallerwebi/dominio/servicios/ServicioBicicletaImpl.java
@@ -1,4 +1,4 @@
-package com.tallerwebi.dominio;
+package com.tallerwebi.dominio.servicios;
 
 import com.tallerwebi.dominio.entidad.Bicicleta;
 import com.tallerwebi.dominio.entidad.EstadoBicicleta;
@@ -7,8 +7,8 @@ import com.tallerwebi.dominio.entidad.Usuario;
 import com.tallerwebi.dominio.excepcion.BicicletaNoDisponible;
 import com.tallerwebi.dominio.excepcion.BicicletaNoEncontrada;
 import com.tallerwebi.dominio.excepcion.BicicletaValidacion;
-import com.tallerwebi.infraestructura.RepositorioBicicleta;
-import com.tallerwebi.presentacion.DatosBicicleta;
+import com.tallerwebi.infraestructura.repositorios.RepositorioBicicleta;
+import com.tallerwebi.presentacion.dto.DatosBicicleta;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
