+++ src/main/java/com/tallerwebi/dominio/servicios/ServicioLogin.java
@@ -1,9 +1,9 @@
-package com.tallerwebi.dominio;
+package com.tallerwebi.dominio.servicios;
 
 import com.tallerwebi.dominio.entidad.Usuario;
 import com.tallerwebi.dominio.excepcion.UsuarioExistente;
 import com.tallerwebi.dominio.excepcion.UsuarioSinRol;
-import com.tallerwebi.presentacion.DatosUsuario;
+import com.tallerwebi.presentacion.dto.DatosUsuario;
 
 public interface ServicioLogin {
     Usuario consultarUsuario(String email, String password);
