+++ src/main/java/com/tallerwebi/presentacion/dto/DatosUsuario.java
@@ -1,15 +1,17 @@
-package com.tallerwebi.presentacion;
+package com.tallerwebi.presentacion.dto;
 
-public class DatosLogin {
+public class DatosUsuario {
     private String email;
     private String password;
+    private String rol;
 
-    public DatosLogin() {
+    public DatosUsuario() {
     }
 
-    public DatosLogin(String email, String password) {
+    public DatosUsuario(String email, String password, String rol) {
         this.email = email;
         this.password = password;
+        this.rol = rol;
     }
 
     public String getEmail() {
@@ -27,5 +29,12 @@ public class DatosLogin {
     public void setPassword(String password) {
         this.password = password;
     }
-}
 
+    public String getRol() {
+        return rol;
+    }
+
+    public void setRol(String rol) {
+        this.rol = rol;
+    }
+}
