+++ src/test/java/com/tallerwebi/integracion/ControladorLoginTest.java
@@ -60,7 +60,7 @@ public class ControladorLoginTest {
 		ModelAndView modelAndView = result.getModelAndView();
         assert modelAndView != null;
         assertThat(modelAndView.getViewName(), equalToIgnoringCase("login"));
-		assertThat(modelAndView.getModel().get("datosLogin").toString(),  containsString("com.tallerwebi.presentacion.DatosLogin"));
+		assertThat(modelAndView.getModel().get("datosLogin").toString(),  containsString("com.tallerwebi.presentacion.dto.DatosLogin"));
 
 	}
 	*/
