+++ src/test/java/com/tallerwebi/presentacion/ControladorBicicletaTest.java
@@ -1,9 +1,11 @@
 package com.tallerwebi.presentacion;
 
-import com.tallerwebi.dominio.ServicioBicicleta;
+import com.tallerwebi.dominio.servicios.ServicioBicicleta;
 import com.tallerwebi.dominio.entidad.Bicicleta;
 import com.tallerwebi.dominio.entidad.Usuario;
 import com.tallerwebi.dominio.excepcion.BicicletaValidacion;
+import com.tallerwebi.presentacion.controladores.ControladorBicicleta;
+import com.tallerwebi.presentacion.dto.DatosBicicleta;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.web.servlet.ModelAndView;
